int->long long intに変更

+++ Molecular_dynamics_cuda.cu
@@ -415,13 +415,13 @@ public:
 int main(void) {
     FPUT_Lattice_1D model = FPUT_Lattice_1D();
 
-    const int output_file_interval = allSteps / 20;
-    const int output_cerr_interval = allSteps / 500;
+    const long long int output_file_interval = allSteps / 20;
+    const long long int output_cerr_interval = allSteps / 500;
     model.settingSize(20, SIZE);
     std::chrono::system_clock::time_point  start, end; // 型は auto で可
     start = std::chrono::system_clock::now(); // 計測開始時間
     model.output_setting();
-    for(int i = 0; i < allSteps;i++) {
+    for(long long int i = 0; i < allSteps;i++) {
         if(i == initialStateSteps)
             model.statistics_reset();
         model.step();
